% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRverse.R
\docType{package}
\name{dartRverse-package}
\alias{dartRverse}
\alias{dartRverse-package}
\title{dartRverse-package}
\description{
The 'dartRverse' combines a set of packages that work together to analyse SNP and silicodart data. 
It is the successor of the package dartR. Due to its high number of dependencies it was necessary to break the package into several, smaller packages to reduce overall dependencies from other packages and to speed up testing.
You can check which packages do exist by typing \code{dartRverse_install()}.
}
\details{
Loading the package attaches all installed dartRverse packages and prints which are installed, which are not, and which are installed but failed to load. Set \code{options(dartRverse.quiet = TRUE)} before \code{library(dartRverse)} to suppress this startup output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/green-striped-gecko/dartRverse}
}

}
\author{
\strong{Maintainer}: Bernd Gruber \email{bernd.gruber@canberra.edu.au}

Authors:
\itemize{
  \item Arthur Georges \email{georges@aerg.edu.au}
  \item Jose L. Mijangos \email{Luis.MijangosAraujo@canberra.edu.au}
  \item Carlo Pacioni \email{carlo.pacioni@delwp.vic.gov.au}
  \item Peter J. Unmack \email{peter.mail2@unmack.net}
  \item Emily Stringer \email{emily.stringer@canberra.edu.au}
  \item Oliver Berry \email{oliver.berry@csiro.au}
}

}
\keyword{internal}
