% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dartRverse_packages}
\alias{dartRverse_packages}
\title{List all packages in the dartRverse}
\usage{
dartRverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include dartRverse in the list?}
}
\description{
List all packages in the dartRverse
}
\examples{
dartRverse_packages()
}
