% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.download.binary.r
\name{gl.download.binary}
\alias{gl.download.binary}
\title{Support function to download binaries from github}
\usage{
gl.download.binary(
  software = NULL,
  os = NULL,
  branch = "main",
  out.dir = tempdir(),
  verbose = 2
)
}
\arguments{
\item{software}{Name of the software package to download, one of "beagle", "colony", "eems", "emibd9", "epos", "faststructure", "gone", "ms", "neestimator", "plink", "popcluster", "stairway2" and "structure". Upper case is converted to lower case. Depending on the software several files will be downloaded. If NULL, a table of the binaries available for each operating system is printed [default NULL].}

\item{os}{The operating system to download the binary for: "windows", "mac" or "linux" (partial matching is allowed). Some binaries are not available for all operating systems. If NULL, the current operating system is used [default NULL].}

\item{branch}{Which branch of the dartRverse github repository to download from, or to list the binaries of. There should be no reason to change it [default "main"].}

\item{out.dir}{The path where to save the binary [default tempdir()].}

\item{verbose}{If 0, download progress and messages are suppressed; any other value prints them [default 2].}
}
\value{
When software is given, invisibly, the path of the folder that holds the unzipped binary (out.dir/software). When software is NULL, NULL (the table of available binaries is printed).
}
\description{
This function supports the download of binaries from github. Those binaries are compiled files that allow to run dartR functions that integrate third party software such as epos (gl.run.epos), NeEstimator (gl.LDNe) or Structure (gl.run.structure). Please be aware this is just to allow for easy install and use of dartR functions. We have explicitly asked the authors of the software for permission to include those binaries here. Understandably some authors prefer to direct users to their own websites to download the software. Here a comment is issued by the function how to do so. Please note: **The third party packages are the work of others and please make the effort to cite them accordingly**. You find the citations under the help pages of the respective functions, for example ?gl.run.epos.
}
\examples{
\dontrun{
# list the binaries available for each operating system
gl.download.binary()
# download PLINK for the current operating system into tempdir()
path <- gl.download.binary("plink")
}
}
