% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRverse.R
\docType{package}
\name{dartRverse-package}
\alias{dartRverse}
\alias{dartRverse-package}
\title{dartRverse-package}
\description{
The 'dartRverse' combines a set of packages that work together to analyse SNP and silicodart data. It is the successor of the previous package dartR. Due to its high number of dependencies it was necessary to break the package into smaller packages to reduce overall dependencies from packages that are no longer maintained on CRAN
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/green-striped-gecko/dartRverse}
}

}
\author{
\strong{Maintainer}: Bernd Gruber \email{bernd.gruber@canberra.edu.au}

Authors:
\itemize{
  \item Arthur Georges \email{georges@aerg.edu.au}
  \item Jose L. Mijangos \email{Luis.MijangosAraujo@canberra.edu.au}
  \item Carlo Pacioni \email{carlo.pacioni@delwp.vic.gov.au}
  \item Peter J. Unmack \email{peter.mail2@unmack.net}
  \item Oliver Berry \email{oliver.berry@csiro.au}
}

Other contributors:
\itemize{
  \item Lindsay V. Clark \email{lvclark@illinois.edu} [contributor]
  \item Floriaan Devloo-Delva \email{Floriaan.Devloo-Delva@csiro.au} [contributor]
  \item Eric Archer \email{eric.archer@noaa.gov} [contributor]
}

}
\keyword{internal}
