% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRverse_install.r
\name{dartRverse_install}
\alias{dartRverse_install}
\title{Supports installation of CRAN and Github package of the dartRverse}
\usage{
dartRverse_install(package = NULL, rep = "CRAN", branch = "main")
}
\arguments{
\item{package}{Name of the package to install, currently [dartR.sim, dartR.spatial, dartR.popgenomics]}

\item{rep}{Which repository is used (CRAN or Github)}

\item{branch}{If Github is used the branch on Github needs to be specified, [either main, beta or dev]}
}
\description{
This functions expects the name of one (or several) dartR packages, the repository (CRAN or Github) and in the case of github the branch (main, dev, beta) to install the identified version of the pacakge. If run with no parameter the current installed packages and their versions are printed
}
\examples{
dartRverse_install()
}
