% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRverse_install.r
\name{dartRverse_install}
\alias{dartRverse_install}
\title{Checks installed package and Supports installation of CRAN and Github packages of the dartRverse}
\usage{
dartRverse_install(
  package = NULL,
  rep = "CRAN",
  branch = "main",
  verbose = TRUE
)
}
\arguments{
\item{package}{Name of the package to install, currently [dartR.base, dartR.data, dartR.sim, dartR.spatial, dartR.popgen, dartR.sexlinked]}

\item{rep}{Which repository is used ('CRAN' or 'Github').}

\item{branch}{If Github is used the branch on Github needs to be specified, [either main, beta or dev]
The 'main' repository on Github is identical with the latest CRAN submission. Important changes and fixes are published under 'beta' and tested there, before the are submitted to CRAN. Hence this might be the best chance 
to look for fixes. All 'dev' branches are 'risky' meaning they have not been tested.}

\item{verbose}{if set to true the current installed packages are printed.}
}
\value{
functions returns NULL
}
\description{
This functions expects the name of one (or several) dartR packages, the repository (CRAN or Github) and in the case of github the branch (main, dev, beta) to install the identified version of the pacakge. If run with no parameter the current installed packages and their versions are printed.
}
\examples{
dartRverse_install()
}
